requirements.txt
mysql-connector-python==9.1.0
python-dotenv==1.0.0
